examples/v1beta1/pipeline/realtimeimageclassfication/components/SERVING/requirements.txt
seldon-core
numpy
boto3
kfserving
cachetools==4.0.0
certifi==2019.11.28
oauthlib==3.1.0
protobuf==3.11.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
tornado==6.0.3
urllib3==1.25.8
websocket-client==0.57.0
gunicorn
Flask==1.1.1
anaconda-client
opencv-contrib-python
imutils
scikit-image
scipy
tensorboard
tensorboard-data-server
tensorboard-plugin-wit
tensorflow==2.5.0
tensorflow-estimator
tesseract
pytesseract
google-cloud-storage==1.26.0
opencv-contrib-python-headless
pillow
cachetools==4.0.0
certifi==2019.11.28
oauthlib==3.1.0
protobuf==3.11.3
rsa==4.0
tornado==6.0.3
urllib3==1.25.8
websocket-client==0.57.0
kubernetes==10.0.1

